blog/backend/requirements.txt
django>=4.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
oss2
gunicorn
